+++ tests/lean-tests/src/specs.rs
@@ -56,3 +56,12 @@ mod issue_1852 {
 #[hax_lib::lean::pure_requires_proof("⟨True, by mvcgen⟩")]
 #[hax_lib::lean::pure_ensures_proof("⟨fun _ => True, by intros; mvcgen⟩")]
 fn custom_pure_proofs(x: u8) {}
+
+/// Resugarings need to be apply also to linked items
+/// https://github.com/cryspen/hax/issues/1945
+mod issue_1945 {
+    #[hax_lib::requires({let x = a; a == 0})]
+    fn mktuple(a: i32) -> bool {
+        {let x = a; a == 0}
+    }
+}